+++ src/shard_shm.h
@@ -161,6 +161,7 @@ attribute_visible SEXP C_shard_segment_write_raw(SEXP seg_ptr, SEXP data, SEXP o
 attribute_visible SEXP C_shard_segment_read_raw(SEXP seg_ptr, SEXP offset, SEXP size);
 attribute_visible SEXP C_shard_segment_gather_read(SEXP seg_ptr, SEXP idx, SEXP type);
 attribute_visible SEXP C_shard_segment_scatter_write(SEXP seg_ptr, SEXP idx, SEXP values, SEXP type);
+attribute_visible SEXP C_shard_segment_read_range(SEXP seg_ptr, SEXP offset, SEXP n_elems, SEXP type);
 attribute_visible SEXP C_shard_segment_protect(SEXP seg_ptr);
 
 /* Wrap an already-open shard_segment_t* in an externalptr with finalizer */